+++ packages/twenty-front/src/modules/ui/display/expandable-list/ExpandableList.tsx
@@ -2,8 +2,8 @@ import React, { ReactElement, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
 import styled from '@emotion/styled';
 
-import { IntersectionObserverWrapper } from '@/activities/components/ExpandableList/IntersectionObserverWrapper';
 import { Chip, ChipVariant } from '@/ui/display/chip/components/Chip';
+import { IntersectionObserverWrapper } from '@/ui/display/expandable-list/IntersectionObserverWrapper';
 import { Dropdown } from '@/ui/layout/dropdown/components/Dropdown';
 
 const StyledContainer = styled.div`
