+++ packages/twenty-front/src/modules/ui/display/expandable-list/IntersectionObserverWrapper.tsx
